fix(filter): guard NaN/±Infinity on min/max_confidence before SQL (#615)
Adds Step 4 finiteness check in handle_filter — non-finite values are
rejected with ERROR_INVALID_PARAMS before query construction, preventing
silent SQLite comparison failures. Splits graph_read_filter_tests.rs
(was 733 lines) into a separate validation_tests module for early-return
tests, keeping both files within the 500-line limit.
